+++ calculator.py
@@ -103,6 +103,7 @@ class sites_opt(object):
                convergence_reached_times=3,
                use_callback_after_step=False,
                exclude_selection=None,
+               freeze_selection=None,
                debug=False):
 
     self.use_callback_after_step = use_callback_after_step
@@ -133,7 +134,11 @@ class sites_opt(object):
     self.exclude_selection = exclude_selection
     self.keep_selection = None
     if self.exclude_selection is not None:
+      assert self.exclude_selection.size() == self.model.size()
       self.keep_selection = ~self.exclude_selection
+    self.freeze_selection = freeze_selection
+    if self.freeze_selection is not None:
+      assert self.freeze_selection.size() == self.model.size()
     self.all_gradients = []
     self.all_targets   = flex.double()
     self.debug = debug
@@ -155,10 +160,16 @@ class sites_opt(object):
     self.number_of_target_and_gradients_calls+=1
     t0=time.time()
     self.set_sites_plus_x()
+    sites_cart = self.sites_plus_x
+    if self.exclude_selection is not None:
+      sites_cart = sites_cart.select(self.keep_selection)
     self.f, self.g = self.restraints_manager.target_and_gradients(
-      sites_cart = self.sites_plus_x)
-    if self.keep_selection is not None:
-      self.g = self.g.set_selected(self.exclude_selection, [0,0,0])
+      sites_cart = sites_cart)
+    if self.exclude_selection is not None:
+      g_full = flex.vec3_double(self.model.size(), [0,0,0])
+      self.g = g_full.set_selected(self.keep_selection, self.g)
+    if self.freeze_selection is not None:
+      self.g = self.g.set_selected(self.freeze_selection, [0,0,0])
     if self.debug:
       self.all_gradients.append(self.g)
       self.all_targets  .append(self.f)